docker環境上で動作確認

+++ scripts/record_windows.sh
@@ -12,9 +12,12 @@ DESKTOPHEIGHT=`xdpyinfo  | grep -oP 'dimensions:\s+\K\S+' | cut -f 2 -d "x"`
 
 
 TMP_DIR=/mnt/tmp
-sudo mkdir -p ${TMP_DIR}
-sudo chmod 777 ${TMP_DIR}
-rm ${TMP_DIR}/*.*
+if [ -e ${TMP_DIR} ]; then
+    rm ${TMP_DIR}/*.*
+else
+    sudo mkdir -p ${TMP_DIR}
+    sudo chmod 777 ${TMP_DIR}
+fi
 
 WLIST=`wmctrl -l`
 
@@ -24,10 +27,10 @@ XID_JUDGE=`echo "$WLIST" | grep 'burger war' | awk '{print $1}'| head -1`
 XID_RQT=`echo "$WLIST" | grep 'rqt_image_view' | awk '{print $1}'| head -1`
 
 
-(gst-launch-1.0 ximagesrc xid=$XID_GAZEBO ! queue ! video/x-raw,framerate=10/1 ! jpegenc ! avimux ! filesink location="${TMP_DIR}/01.avi")&
-(gst-launch-1.0 ximagesrc xid=$XID_JUDGE ! queue ! video/x-raw,framerate=1/1 ! jpegenc ! avimux ! filesink location="${TMP_DIR}/02.avi")&
-(gst-launch-1.0 ximagesrc xid=$XID_RQT ! video/x-raw,framerate=10/1 ! jpegenc ! avimux ! filesink location="${TMP_DIR}/03.avi")&
+(gst-launch-1.0 ximagesrc xid=$XID_GAZEBO show-pointer=false ! queue ! video/x-raw,framerate=10/1 ! jpegenc ! avimux ! filesink location="${TMP_DIR}/01.avi")&
+(gst-launch-1.0 ximagesrc xid=$XID_JUDGE show-pointer=false ! queue ! video/x-raw,framerate=1/1 ! jpegenc ! avimux ! filesink location="${TMP_DIR}/02.avi")&
+(gst-launch-1.0 ximagesrc xid=$XID_RQT show-pointer=false ! video/x-raw,framerate=10/1 ! jpegenc ! avimux ! filesink location="${TMP_DIR}/03.avi")&
 
-(gst-launch-1.0 ximagesrc ! queue ! video/x-raw,framerate=10/1 ! jpegenc ! avimux ! filesink location="${TMP_DIR}/00.avi")&
+(gst-launch-1.0 ximagesrc show-pointer=false ! queue ! video/x-raw,framerate=10/1 ! jpegenc ! avimux ! filesink location="${TMP_DIR}/00.avi")&
